+++ Sources/ApolloTestSupport/MockObject.swift
@@ -1,0 +1,14 @@
+import ApolloAPI
+
+public class Mock<O: Object> {
+  public var data: JSONObject
+
+  public init() {
+    data = ["__typename": O.__typename.description]
+  }
+
+}
+
+public protocol Mockable {
+  associatedtype MockFields
+}
